+++ packages/http-server/src/types.ts
@@ -101,7 +101,10 @@ export interface HttpServerModuleOptions {
 	port?: number | string;
 	contextFactory?: ContextFactory<unknown>;
 	validationFactory?: ValidationFactory;
-	errorsHandling?: {
+	errorHandling?: {
+		/**
+		 * enabled by default. False if process.env.NODE_ENV === 'production'
+		 */
 		exposeStack?: boolean;
 	};
 	querystringJsonParsing?: boolean;