+++ packages/health-checks/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.21](https://github.com/HPInc/davinci/compare/@davinci/health-checks@2.1.20...@davinci/health-checks@2.1.21) (2023-01-17)
+
+**Note:** Version bump only for package @davinci/health-checks
+
+
+
+
+
+## [2.1.20](https://github.com/HPInc/davinci/compare/@davinci/health-checks@2.1.19...@davinci/health-checks@2.1.20) (2023-01-17)
+
+**Note:** Version bump only for package @davinci/health-checks
+
+
+
+
+
+## [2.1.19](https://github.com/HPInc/davinci/compare/@davinci/health-checks@2.1.18...@davinci/health-checks@2.1.19) (2023-01-16)
+
+**Note:** Version bump only for package @davinci/health-checks
+
+
+
+
+
+## [2.1.18](https://github.com/HPInc/davinci/compare/@davinci/health-checks@2.1.17...@davinci/health-checks@2.1.18) (2023-01-16)
+
+**Note:** Version bump only for package @davinci/health-checks
+
+
+
+
+
+## [2.1.17](https://github.com/HPInc/davinci/compare/@davinci/health-checks@2.1.16...@davinci/health-checks@2.1.17) (2023-01-16)
+
+**Note:** Version bump only for package @davinci/health-checks
+
+
+
+
+
 ## [2.1.16](https://github.com/HPInc/davinci/compare/@davinci/health-checks@2.1.15...@davinci/health-checks@2.1.16) (2023-01-04)
 
 **Note:** Version bump only for package @davinci/health-checks