+++ packages/messaging-amqp/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.12](https://github.com/HPInc/davinci/compare/@davinci/messaging-amqp@2.1.11...@davinci/messaging-amqp@2.1.12) (2023-01-16)
+
+**Note:** Version bump only for package @davinci/messaging-amqp
+
+
+
+
+
 ## [2.1.11](https://github.com/HPInc/davinci/compare/@davinci/messaging-amqp@2.1.10...@davinci/messaging-amqp@2.1.11) (2023-01-04)
 
 **Note:** Version bump only for package @davinci/messaging-amqp